feat: add emmc driver

+++ inc/emmc.h
@@ -43,7 +43,7 @@ struct emmc {
 };
 
 int emmc_init(struct emmc *self);
-int emmc_read(struct emmc *self, void *buf, size_t cnt);
-int emmc_write(struct emmc *self, void *buf, size_t cnt);
+size_t emmc_read(struct emmc *self, void *buf, size_t cnt);
+size_t emmc_write(struct emmc *self, void *buf, size_t cnt);
 uint64_t emmc_seek(struct emmc *self, uint64_t off);
